update models from sheets following sckan qc
in general there have been many changes made in order to bring the
apinatomy model representations in line with their respective
diagrams (i.e. docs/{model-name}.svg

changes made for general consistency

1. all ontologyTerms for anatomical parts have been updated to use
ILX: or UBERON: terms, use of any other prefix is a bug and should be
reported as such

2. processes passing through white and gray rami also pass through the
corresponding sympathetic chain ganglion

3. labels for dend-{bag,tube} have been update to reflect that the
dendrite lyph templates are currently used in a way that is ambiguous
to refer to both traditional dendrites and structures that are axons
with sensory endings, a future update will disambiguate these cases
and also address issues in the peripheral nervous system where an
absense of explicit dendrites was incorrectly interpreted as the
presenece of a dendrite (the default assumption is different from the
central nervous system)

4. in most, but not all cases we have switched to using levelTargets
to model neuron populations with somas in multiple locations that send
processes through the same structures e.g. kblad 4/8, aacar 5

5. in cases where a housing region was unknown we selected something
like a least upper bound to constrain the location e.g. for axons
projecting into spinal white matter from the periphery we listed
central nervous system as a target, and for sensory processes whose
anatomical original was unknown we listed multicellular organis aka
whole body

6. where more specific ontology terms are available for specific lyphs
and layers we attempt to use them e.g. myenteric plexus of stomach
instead of the more general myenteric plexus

individual changes are too numerous to list but nearly all have been
documented specifically as part of the qc process, ping for details if
needed, some general patterns for issues that were correct follow

1. inversion of layer ordering, e.g. lyphs specified inside out so
that e.g. the colon was inside out with the serosa as the inner layer
and the lumen the outer layer, or e.g. blood specified as being
outside the blood vessel wall

2. frame shifts, insertions, etc. corresponding to biological genomic
mutations, induced by copy/paste errors of various kinds, nearly all
leading to off by 1 errors or incorrect aliasing, or sometimes
incorrect typing of neurons with similar structure

3. axon/dendrite confusion independent of axon sensory/dendrite issues

there are a number of cases where the notation used in apinatomy
diagrams can be ambiguous, some general categories where a mismatch
should be expected are as follows

1. branching of neurons into the same structure, in nearly all cases
where a diagram represents multiple lyphs that pass through or
terminate in a single housing lyph only a single branch is
represented, e.g. sstom {11..14}, aacar {4,6}, in many of these cases
the branching is induced by projections onto multiple different cell
populations and could be handled during model expansion

2. neurons passing into and out of an internal lyph back into the
parent, if a branch happens inside the internal lyph the notation is
equivalent to a branch occuring in the housing lyph prior to entering
or even after exiting the internal lyph, e.g. pancr {2,4}

3. the location of implicit nodes in cases where continuous lines in a
diagram could be in one or more locations (related to the internal vs
housing lyph boundaries issue mentioned above), see e.g. aacar
{4,6,11,13}, digrams can and do vary greatly from models in such cases

notational confusion in apinatomy diagrams

1. orthognoal crossing lines of the same color that are numbered but
otherwise appear continuous, we have fixed most of these cases, but
every single person who has ever looked at the kblad diagram has
confused populations 2 and 4 (hopefully the recent changes are
sufficient to disambiguate the two), this issue is not unique to
apinatomy diagram notation, anyone reading a diagram should still take
care to check whether additional notation exists when an orthognal
line crossing of lines is encountered

2. implicit dendrites, the default assumption switches across the
cns/pns boundary

3. lack of indication of the axis of rotation for lyphs, this is
hypothesized to be the primary cause of the layer order inversion
issues that we have encountered, some form of asymmetry is critical
for quick an accurate reading of diagrams, cysts and bags can usually
be read without issue due to their inherent asymmetry, but tubes need
explicit axis, readers encountering a tube without specification
should take special care and also report the issue

4. use of a scaffold region by itself without a corresponding lyph
from a connectivity model, most of these have been fixed, but some
cases may not have been corrected, this is particularly confusing when
the region is empty and used as a reminder that there is something in
there that a process passes through, but where the exact location is
unknown, in such cases the region should be left out, but we often
instinctively want to close the world and include it in the
connectivity diagram, thus explicit lyphs in such regions are critical
for clear communication

5. implicit boundaries between regions, e.g. kblad {6,7}, bromo 1,
sdcol b, these have been almost entirely resolved in new diagrams by
keeping even regions that are contiguous separate

6. the converse of 5 seen in e.g. pancr 4 where proximal duodenum and
gastric antrum are intended to be continuous but there is not an
explicit link/node that denotes this in the diagram

7. chains drawn as going directly to an intermediate layer of a lyph
without explicitly passing through the overlying/underlying layers, in
all cases the chain must pass through the overlying/underlying lyph,
but there is not bottom up way to know whether the outer or inner
layer is the entry point, in most existing cases e.g. kblad {2,3} it
is the outer layer

another general trend that we have observed is that nearly all errors
that we found in the ApiNATOMY models after initial checking against
NPO were things that NPO could never have caught because they were
e.g. incorrect identifiers for layers that the current version of NPO
does not encode, the lyph layer inversions that literally no one
caught because most tools do not care whether a sequece of housing
lyphs is provided from inside to outside or outside to inside so long
as their identifiers appear in the expected order (invariance to
topology is an interesting blindspot in existing systems)